// packages/reflection/tests/fixtures/fixture/v1/service_pb.ts
// @generated by protoc-gen-es v2.16.0 with parameter "target=ts,import_extension=.ts,erasable_syntax=true"
// @generated from file fixture/v1/service.proto (package fixture.v1, syntax proto3)
/* eslint-disable */

import type { Message } from "@bufbuild/protobuf";
import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv2";
import type { EmptySchema } from "@bufbuild/protobuf/wkt";
import { file_google_protobuf_empty } from "@bufbuild/protobuf/wkt";
import type { MetaSchema } from "./common_pb.ts";
import { file_fixture_v1_common } from "./common_pb.ts";
import type { Extendable } from "./legacy_pb.ts";
import { file_fixture_v1_legacy } from "./legacy_pb.ts";
import { file_fixture_v1_options } from "./options_pb.ts";

/**
 * Describes the file fixture/v1/service.proto.
 */
export const file_fixture_v1_service: GenFile =
    /*@__PURE__*/
    fileDesc(
        "ChhmaXh0dXJlL3YxL3NlcnZpY2UucHJvdG8SCmZpeHR1cmUudjEiXAoKR2V0UmVxdWVzdBIMCgJpZBgBIAEoCUgAEhAKBm51bWJlchgCIAEoA0gAEicKB3BheWxvYWQYAyABKAsyFi5maXh0dXJlLnYxLkV4dGVuZGFibGVCBQoDa2V5MoMBCg5GaXh0dXJlU2VydmljZRI5CgNHZXQSFi5maXh0dXJlLnYxLkdldFJlcXVlc3QaEC5maXh0dXJlLnYxLk1ldGEiCIq1GARyZWFkEjYKBFBpbmcSFi5nb29nbGUucHJvdG9idWYuRW1wdHkaFi5nb29nbGUucHJvdG9idWYuRW1wdHliBnByb3RvMw",
        [file_fixture_v1_common, file_fixture_v1_legacy, file_fixture_v1_options, file_google_protobuf_empty],
    );

/**
 * @generated from message fixture.v1.GetRequest
 */
export type GetRequest = Message<"fixture.v1.GetRequest"> & {
    /**
     * @generated from oneof fixture.v1.GetRequest.key
     */
    key:
        | {
              /**
               * @generated from field: string id = 1;
               */
              value: string;
              case: "id";
          }
        | {
              /**
               * @generated from field: int64 number = 2;
               */
              value: bigint;
              case: "number";
          }
        | { case: undefined; value?: undefined };

    /**
     * @generated from field: fixture.v1.Extendable payload = 3;
     */
    payload?: Extendable | undefined;
};

/**
 * Describes the message fixture.v1.GetRequest.
 * Use `create(GetRequestSchema)` to create a new message.
 */
export const GetRequestSchema: GenMessage<GetRequest> = /*@__PURE__*/ messageDesc(file_fixture_v1_service, 0);

/**
 * @generated from service fixture.v1.FixtureService
 */
export const FixtureService: GenService<{
    /**
     * @generated from rpc fixture.v1.FixtureService.Get
     */
    get: {
        methodKind: "unary";
        input: typeof GetRequestSchema;
        output: typeof MetaSchema;
    };
    /**
     * @generated from rpc fixture.v1.FixtureService.Ping
     */
    ping: {
        methodKind: "unary";
        input: typeof EmptySchema;
        output: typeof EmptySchema;
    };
}> = /*@__PURE__*/ serviceDesc(file_fixture_v1_service, 0);